python/cli_flask/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
applicationinsights==0.11.10
attrs==25.3.0
backports.csv==1.0.7
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
cli-helpers==0.2.3
click==8.2.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
Deprecated==1.2.18
enum34==1.1.10
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
future==1.0.0
greenlet==3.2.3
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
multipledispatch==1.0.0
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt-toolkit==2.0.10
Pygments==2.19.2
pykrx==1.0.51
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.42
sqlparse==0.2.4
tabulate==0.9.0
terminaltables==3.1.10
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.2
zope.interface==7.2
